+++ metadata.rb
@@ -4,4 +4,4 @@ maintainer_email 'd.librera@cantierecreativo.net'
 license          'All rights reserved'
 description      'Installs/Configures certbot'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.2.0'
+version          '0.2.1'